rename cmt to cachemanager
